% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-data-format.R
\name{check_data_format}
\alias{check_data_format}
\title{Check Data follow Required Format as per the Template}
\usage{
check_data_format(..., template, complete = FALSE, join = FALSE)
}
\arguments{
\item{...}{A list of named data frames of the data}

\item{template}{A list of named data frames that make up the template}

\item{complete}{A logical indicating if all tables present in the template
need to be supplied in the ... argument. If TRUE all tables need to be
provided in ... argument, default is set to FALSE.}

\item{join}{A logical indicating if joins between the tables should be
checked. If TRUE joins are checked, default is set to FALSE.}
}
\value{
A list of named data frames of the data
}
\description{
Check the types, ranges, sets of values, primary keys, missing values, unique
and join requirements of the template against the data to ensure the data
follows the format outlined in the template.
}
\details{
The template argument should contain all the sheets in the template
while the ... argument can be which ever set of data tables you want to check
against the template. This means either one, several or all the data tables
can be checked against the template.

If complete is set to TRUE then all the data tables need to be supplied.
If join is set to TRUE then the joins between the provided tables are
checked.
}
\examples{
\dontrun{
check_data_format(
  outing = outing,
  template = demo_template_fish_exploit,
  complete = FALSE,
  join = FALSE
)

check_data_format(
  site = site,
  outing = outing,
  capture = capture,
  recapture = recapture,
  template = demo_template_fish_exploit,
  complete = TRUE,
  join = TRUE
)
}
}
