% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_template}
\alias{check_template}
\title{Check that template is valid.}
\usage{
check_template(x)
}
\arguments{
\item{x}{A data.frame of the template.}
}
\value{
An invisible data.frame.
}
\description{
Check that template is valid.
}
