% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{chk_to_type}
\alias{chk_to_type}
\title{Convert chk to type}
\usage{
chk_to_type(x)
}
\arguments{
\item{x}{A vector of the check values (e.g. c(1, 2, NA)).}
}
\value{
A character string of the type.
}
\description{
Convert chk to type
}
