% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_convert.R
\name{chkrow_to_type}
\alias{chkrow_to_type}
\title{Convert chk row to character vector of types}
\usage{
chkrow_to_type(x)
}
\arguments{
\item{x}{A data.frame with chk values for each column.}
}
\value{
A character string of the types.
}
\description{
Convert chk row to character vector of types
}
