% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_template_count}
\alias{demo_template_count}
\title{Multiple Table Demo Template Count}
\format{
A named list of data.frames:
\describe{
\item{visit}{visit table}
\item{crew}{crew table}
\item{count}{count table}
...
}
}
\usage{
demo_template_count
}
\description{
Multiple Table Demo Template Count
}
\keyword{datasets}
