% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_template_data_entry}
\alias{demo_template_data_entry}
\title{Demo template based on data entry upload}
\format{
An object of class \code{data.frame} with 5 rows and 6 columns.
}
\usage{
demo_template_data_entry
}
\description{
Demo template based on data entry upload
}
\keyword{datasets}
