% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_template_fish_exploit}
\alias{demo_template_fish_exploit}
\title{Multiple Table Demo Template Fish Exploit}
\format{
A named list of data.frames:
\describe{
\item{site}{site table}
\item{outing}{outing table}
\item{capture}{capture table}
\item{recapture}{recapture table}
...
}
}
\usage{
demo_template_fish_exploit
}
\description{
Multiple Table Demo Template Fish Exploit
}
\keyword{datasets}
