% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_template_logger}
\alias{demo_template_logger}
\title{Demo template based on data entry upload}
\format{
An object of class \code{data.frame} with 1 rows and 3 columns.
}
\usage{
demo_template_logger
}
\description{
Demo template based on data entry upload
}
\keyword{datasets}
