% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_template_points_historic}
\alias{demo_template_points_historic}
\title{Demo template for points in the historic data module}
\format{
An object of class \code{data.frame} with 1 rows and 4 columns.
}
\usage{
demo_template_points_historic
}
\description{
Demo template for points in the historic data module
}
\keyword{datasets}
