% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_template_tag}
\alias{demo_template_tag}
\title{Single Demo Template}
\format{
A data.frame with columns:
\describe{
\item{name}{Row description}
\item{date}{Date the fish was caught}
\item{tag_1}{The number from the first tag}
\item{tag_2}{The number from the second tag}
\item{harvested}{Either yes or no for if the fish was harvested}
\item{comment}{Comments about the capture}
}
}
\usage{
demo_template_tag
}
\description{
Sample template for one table.
}
\keyword{datasets}
