% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template-human.R
\name{template_human}
\alias{template_human}
\title{Create Human Readable Template}
\usage{
template_human(template)
}
\arguments{
\item{template}{A data.frame of the template. with row names including
'description', 'chk', 'unique' and optionally, 'example'.}
}
\value{
The modified data.frame.
}
\description{
Convert code written template to a human readable version.
}
\details{
The template must contain a column called 'name'. The name column
includes rows named 'description', 'chk', 'unique' and optionally,
'example', 'unique' and 'join'.
}
\examples{
template_human(demo_template_count$visit)
template_human(demo_template_fish_exploit$capture)
}
