% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{template_modify}
\alias{template_modify}
\title{Modify template}
\usage{
template_modify(template)
}
\arguments{
\item{template}{A data.frame of the template with row names including 'description', 'chk', 'unique' and optionally, 'pkey' and 'example'}
}
\value{
The modified data.frame.
}
\description{
Modify template to human-readable format.
}
