% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-data-format.R
\name{check_data_format}
\alias{check_data_format}
\title{Check Data follow Required Format as per the Template}
\usage{
check_data_format(..., template, joins = NULL, complete = FALSE)
}
\arguments{
\item{...}{A list of named data frames of the data}

\item{template}{A list of named data frames that make up the template}

\item{joins}{NULL I think this can be removed...}

\item{complete}{A logical indicating if all tables present in the template
need to be supplied in the ... argument. If TRUE all tables need to be
provided in ... argument, default is set to FALSE. This must be set to TRUE
to check joins between the tables.}
}
\value{
A list of named data frames of the data
}
\description{
Check the types, ranges, sets of values, primary keys, missing values, unique
and join requirements of the template against the data to ensure the data
follows the format outlined in the template.
}
\details{
The template argument should contain all the sheets in the template
while the ... argument can be which ever set of data tables you want to check
against the template. This means either one, several or all the data tables
can be checked against the template.

If complete is set to TRUE then all the data tables need to be supplied and
the function can check joins between the tables.
}
\examples{
\dontrun{
check_data_format(
  outing = outing,
  template = demo_template_fish_exploit,
  complete = FALSE
)

check_data_format(
  outing = outing,
  capture = capture,
  recapture = recapture,
  template = demo_template_fish_exploit,
  complete = TRUE
)
}
}
